add support for rewriting Arc<[T]> into Arc<Box<[T]>> for more fixable scenarios
