+++ clippy_lints/src/vec_to_rc_slice.rs
@@ -1,12 +1,13 @@
 use clippy_utils::diagnostics::span_lint_and_then;
-use clippy_utils::res::{MaybeDef, MaybeTypeckRes};
+use clippy_utils::res::{MaybeDef, MaybeResPath, MaybeTypeckRes};
 use clippy_utils::source::snippet_with_context;
-use clippy_utils::sym;
+use clippy_utils::{qpath_generic_tys, sym};
 use rustc_errors::Applicability;
-use rustc_hir::{Expr, ExprKind};
+use rustc_hir::{Expr, ExprKind, Node, TyKind};
 use rustc_lint::{LateContext, LateLintPass};
 use rustc_middle::ty::{self, Ty};
 use rustc_session::declare_lint_pass;
+use rustc_span::Span;
 
 declare_clippy_lint! {
     /// ### What it does
@@ -71,6 +72,23 @@ fn rc_slice_wrapper(cx: &LateContext<'_>, ty: Ty<'_>) -> Option<&'static str> {
     }
 }
 
+/// If the expression is the init of a `let` statement with a type annotation like
+/// `Arc<[T]>` or `Rc<[T]>`, returns the span of the inner slice type (e.g. `[T]`).
+fn let_ty_inner_slice_span(cx: &LateContext<'_>, expr: &Expr<'_>) -> Option<Span> {
+    if let Node::LetStmt(let_stmt) = cx.tcx.parent_hir_node(expr.hir_id)
+        && let Some(hir_ty) = let_stmt.ty
+        && let TyKind::Path(ref qpath) = hir_ty.kind
+        && let Some(def_id) = hir_ty.basic_res().opt_def_id()
+        && matches!(cx.tcx.get_diagnostic_name(def_id), Some(sym::Arc | sym::Rc))
+        && let Some(inner_ty) = qpath_generic_tys(qpath).next()
+        && matches!(inner_ty.kind, TyKind::Slice(_))
+    {
+        Some(inner_ty.span)
+    } else {
+        None
+    }
+}
+
 impl<'tcx> LateLintPass<'tcx> for VecToRcSlice {
     fn check_expr(&mut self, cx: &LateContext<'tcx>, e: &'tcx Expr<'_>) {
         if e.span.from_expansion() {
@@ -109,19 +127,28 @@ impl<'tcx> LateLintPass<'tcx> for VecToRcSlice {
 fn emit_lint(cx: &LateContext<'_>, expr: &Expr<'_>, vec_expr: &Expr<'_>, wrapper: &str) {
     let mut app = Applicability::MaybeIncorrect;
     let vec_snippet = snippet_with_context(cx, vec_expr.span, expr.span.ctxt(), "<vec>", &mut app).0;
+    let inner_slice_span = let_ty_inner_slice_span(cx, expr);
+
     span_lint_and_then(
         cx,
         VEC_TO_RC_SLICE,
         expr.span,
         format!("converting a `Vec<T>` to `{wrapper}<[T]>` copies all elements to a new allocation"),
         |diag| {
-            diag.span_suggestion(
-                expr.span,
-                "use `into_boxed_slice()` to avoid the copy",
-                format!("{wrapper}::new({vec_snippet}.into_boxed_slice())"),
-                app,
-            );
-            diag.note(format!("this gives `{wrapper}<Box<[T]>>` and avoids the copy"));
+            let expr_sugg = format!("{wrapper}::new({vec_snippet}.into_boxed_slice())");
+
+            if let Some(ty_span) = inner_slice_span {
+                let mut ty_app = Applicability::MachineApplicable;
+                let slice_snippet = snippet_with_context(cx, ty_span, ty_span.ctxt(), "_", &mut ty_app).0;
+                diag.multipart_suggestion(
+                    "use `into_boxed_slice()` to avoid the copy",
+                    vec![(expr.span, expr_sugg), (ty_span, format!("Box<{slice_snippet}>"))],
+                    app,
+                );
+            } else {
+                diag.span_suggestion(expr.span, "use `into_boxed_slice()` to avoid the copy", expr_sugg, app);
+                diag.note(format!("this gives `{wrapper}<Box<[T]>>` and avoids the copy"));
+            }
         },
     );
 }
