+++ tests/ui/vec_to_rc_slice.fixed
@@ -4,9 +4,39 @@ use std::rc::Rc;
 use std::sync::Arc;
 
 fn main() {
+    // Should lint: Vec<T>.into() -> Arc<[T]>
+    let v: Vec<u8> = vec![1, 2, 3];
+    let _a: Arc<Box<[u8]>> = Arc::new(v.into_boxed_slice());
+    //~^ vec_to_rc_slice
+
+    // Should lint: Arc::from(vec)
+    let v: Vec<u8> = vec![1, 2, 3];
+    let _a: Arc<Box<[u8]>> = Arc::new(v.into_boxed_slice());
+    //~^ vec_to_rc_slice
+
+    // Should lint: From::from(vec) when target is Arc<[T]>
+    let v: Vec<u8> = vec![1, 2, 3];
+    let _a: Arc<Box<[u8]>> = Arc::new(v.into_boxed_slice());
+    //~^ vec_to_rc_slice
+
     // Should lint: fully qualified Arc from
     let v: Vec<u8> = vec![1, 2, 3];
     let _a = Arc::new(v.into_boxed_slice());
+    //~^ vec_to_rc_slice
+
+    // Should lint: Vec<T>.into() -> Rc<[T]>
+    let v: Vec<u8> = vec![1, 2, 3];
+    let _a: Rc<Box<[u8]>> = Rc::new(v.into_boxed_slice());
+    //~^ vec_to_rc_slice
+
+    // Should lint: Rc::from(vec)
+    let v: Vec<u8> = vec![1, 2, 3];
+    let _a: Rc<Box<[u8]>> = Rc::new(v.into_boxed_slice());
+    //~^ vec_to_rc_slice
+
+    // Should lint: From::from(vec) when target is Rc<[T]>
+    let v: Vec<u8> = vec![1, 2, 3];
+    let _a: Rc<Box<[u8]>> = Rc::new(v.into_boxed_slice());
     //~^ vec_to_rc_slice
 
     // Should lint: fully qualified Rc from
