+++ tests/ui/vec_to_rc_slice_unfixable.rs
@@ -8,36 +8,12 @@ fn accept_arc_slice(_: Arc<[u8]>) {}
 fn accept_rc_slice(_: Rc<[u8]>) {}
 
 fn main() {
-    // Type annotation constrains the result type; fix changes it
-    let v: Vec<u8> = vec![1, 2, 3];
-    let _a: Arc<[u8]> = v.into();
-    //~^ vec_to_rc_slice
-
-    let v: Vec<u8> = vec![1, 2, 3];
-    let _a: Arc<[u8]> = Arc::from(v);
-    //~^ vec_to_rc_slice
-
-    let v: Vec<u8> = vec![1, 2, 3];
-    let _a: Arc<[u8]> = From::from(v);
-    //~^ vec_to_rc_slice
-
-    let v: Vec<u8> = vec![1, 2, 3];
-    let _a: Rc<[u8]> = v.into();
-    //~^ vec_to_rc_slice
-
-    let v: Vec<u8> = vec![1, 2, 3];
-    let _a: Rc<[u8]> = Rc::from(v);
-    //~^ vec_to_rc_slice
-
-    let v: Vec<u8> = vec![1, 2, 3];
-    let _a: Rc<[u8]> = From::from(v);
-    //~^ vec_to_rc_slice
-
-    // Function signature constrains the result type; fix requires downstream changes
+    // Should lint: result passed to a function expecting Arc<[T]> (fix requires downstream changes)
     let v: Vec<u8> = vec![1, 2, 3];
     accept_arc_slice(v.into());
     //~^ vec_to_rc_slice
 
+    // Should lint: result passed to a function expecting Rc<[T]> (fix requires downstream changes)
     let v: Vec<u8> = vec![1, 2, 3];
     accept_rc_slice(Rc::from(v));
     //~^ vec_to_rc_slice
