+++ tests/ui/vec_to_rc_slice_unfixable.stderr
@@ -1,68 +1,20 @@
 error: converting a `Vec<T>` to `Arc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:13:25
-   |
-LL |     let _a: Arc<[u8]> = v.into();
-   |                         ^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Arc::new(v.into_boxed_slice())`
-   |
-   = note: this gives `Arc<Box<[T]>>` and avoids the copy
-   = note: `-D clippy::vec-to-rc-slice` implied by `-D warnings`
-   = help: to override `-D warnings` add `#[allow(clippy::vec_to_rc_slice)]`
-
-error: converting a `Vec<T>` to `Arc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:17:25
-   |
-LL |     let _a: Arc<[u8]> = Arc::from(v);
-   |                         ^^^^^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Arc::new(v.into_boxed_slice())`
-   |
-   = note: this gives `Arc<Box<[T]>>` and avoids the copy
-
-error: converting a `Vec<T>` to `Arc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:21:25
-   |
-LL |     let _a: Arc<[u8]> = From::from(v);
-   |                         ^^^^^^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Arc::new(v.into_boxed_slice())`
-   |
-   = note: this gives `Arc<Box<[T]>>` and avoids the copy
-
-error: converting a `Vec<T>` to `Rc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:25:24
-   |
-LL |     let _a: Rc<[u8]> = v.into();
-   |                        ^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Rc::new(v.into_boxed_slice())`
-   |
-   = note: this gives `Rc<Box<[T]>>` and avoids the copy
-
-error: converting a `Vec<T>` to `Rc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:29:24
-   |
-LL |     let _a: Rc<[u8]> = Rc::from(v);
-   |                        ^^^^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Rc::new(v.into_boxed_slice())`
-   |
-   = note: this gives `Rc<Box<[T]>>` and avoids the copy
-
-error: converting a `Vec<T>` to `Rc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:33:24
-   |
-LL |     let _a: Rc<[u8]> = From::from(v);
-   |                        ^^^^^^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Rc::new(v.into_boxed_slice())`
-   |
-   = note: this gives `Rc<Box<[T]>>` and avoids the copy
-
-error: converting a `Vec<T>` to `Arc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:38:22
+  --> tests/ui/vec_to_rc_slice_unfixable.rs:13:22
    |
 LL |     accept_arc_slice(v.into());
    |                      ^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Arc::new(v.into_boxed_slice())`
    |
    = note: this gives `Arc<Box<[T]>>` and avoids the copy
+   = note: `-D clippy::vec-to-rc-slice` implied by `-D warnings`
+   = help: to override `-D warnings` add `#[allow(clippy::vec_to_rc_slice)]`
 
 error: converting a `Vec<T>` to `Rc<[T]>` copies all elements to a new allocation
-  --> tests/ui/vec_to_rc_slice_unfixable.rs:42:21
+  --> tests/ui/vec_to_rc_slice_unfixable.rs:18:21
    |
 LL |     accept_rc_slice(Rc::from(v));
    |                     ^^^^^^^^^^^ help: use `into_boxed_slice()` to avoid the copy: `Rc::new(v.into_boxed_slice())`
    |
    = note: this gives `Rc<Box<[T]>>` and avoids the copy
 
-error: aborting due to 8 previous errors
+error: aborting due to 2 previous errors
